refactor: rename replay policy to replay mode and update related documentation
- Renamed `ReplayPolicy` to `ReplayMode` across the codebase for clarity and consistency.
- Updated references in multiple files, including AGENTS.md, IDEAS.md, and various source files, to reflect the new naming convention.
- Introduced a new `replay-mode.ts` file to encapsulate replay mode logic and types, enhancing modularity.
- Revised documentation to clarify the behavior of replay modes and their impact on game analysis processes.
