+++ src/replay/index.ts
@@ -1,2 +1,2 @@
-export type { ReplayMode } from '#replay/replay-policy';
-export { resolveReplayMode, resolveEffectiveReplayMode } from '#replay/replay-policy';
+export type { ReplayMode } from '#replay/replay-mode';
+export { resolveReplayMode, resolveEffectiveReplayMode } from '#replay/replay-mode';
